requirements.txt
streamlit~=1.34.0
google-generativeai
python-dotenv~=1.0.1
langchain~=0.2.1
PyPDF2~=3.0.1
chromadb
faiss-cpu
langchain_google_genai